+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 
 v1.3.1 (2022.08.xx)
 -------------------
+- Add variants to media
 - Bump Kotlin 1.7.10
 
 v1.3.0 (2022.07.05)